worked on server client stuff. also GameWorld

+++ core/src/sophomoreproject/game/networking/Account.java
@@ -5,9 +5,9 @@ import java.io.Serializable;
 public class Account implements Serializable {
     public String username;
     public String password;
-    public long accountID;
+    public int accountID;
 
-    public Account(String username, String password, long accountID) {
+    public Account(String username, String password, int accountID) {
         this.username = username;
         this.password = password;
         this.accountID = accountID;
